+++ ICD.Connect.Audio.QSys/Controls/NamedControlsVolumeDevice.cs
@@ -6,8 +6,9 @@ using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Services.Logging;
 using ICD.Common.Utils.Xml;
 using ICD.Connect.API.Nodes;
+using ICD.Connect.Audio.Controls.Mute;
+using ICD.Connect.Audio.Controls.Volume;
 using ICD.Connect.Audio.QSys.CoreControls;
-using ICD.Connect.Audio.Controls;
 using ICD.Connect.Audio.QSys.CoreControls.NamedControls;
 using ICD.Connect.Audio.Repeaters;