+++ ICD.Connect.Audio/Controls/Volume/AbstractVolumeLevelDeviceControl.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using ICD.Connect.API.Commands;
 using ICD.Connect.API.Nodes;
-using ICD.Connect.Audio.Console;
+using ICD.Connect.Audio.Console.Volume;
 using ICD.Connect.Audio.Repeaters;
 using ICD.Connect.Devices;
 
-namespace ICD.Connect.Audio.Controls
+namespace ICD.Connect.Audio.Controls.Volume
 {
 	public abstract class AbstractVolumeLevelDeviceControl<T> : AbstractVolumePositionDeviceControl<T>, IVolumeLevelDeviceControl
 		where T : IDeviceBase