+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added microphone control interface and abstraction
+ - Added Volume Feedback Log
+ 
+### Changed
+ - Moved controls into subdirectories/namespaces
+ - Inserted AbstractVolumeDeviceControl into volume control heirarchy
 
 ## [8.2.0] - 2018-11-20
 ### Added